discussion with the borrow checker
The two lifetime arguments on BorrowName were getting troublesome.
It seemed simpler to eliminate one of them by making BorrowName own
the label index. This means that the BorrowName itself shouldn't be
copied around, but we should refer to it instead. The exception is
when it is constructed: I assume that a return value optimization will
deal with that.