start dnsname revamp
Most methods are moving into the DnsName trait.

We now have a notion of a work pad, which is incrementally initialized
while parsing a name. The aim is to avoid redundant work.

Use associated types in the DnsName trait to own the awkward
lifetimes, instead of making them generic parameters. This leads to
some complicated trait bounds on the methods with default
implementations, but it is much easier to explain to the borrow
checker what is going on (e.g. no PhantomData in the iterator).