drop DnsLabels
On balance it seems very complicated to be able to refer to a
zero-copy name inside a DNS message.

I originally wanted this so that I could create a trie lookup key
directly from the wire, without going via an intermediate copy of the
name; but this isn't as useful as I wanted because at the end of the
lookup we need to compare the HeapName from the trie with the
name from the wire, which is easier if it has been canonicalized.

(There remains a question whether it's faster to canonicalize then
compare, or compare directly, and I have just made it harder to answer
this question...)